+++ lib/kitchen/directions/move_solutions_to_answer_key/v1.rb
@@ -9,6 +9,8 @@ module Kitchen::Directions::MoveSolutionsToAnswerKey
           Strategies::Calculus
         when :uphysics
           Strategies::UPhysics
+        when :american_government
+          Strategies::AmericanGovernment
         else
           raise 'No such strategy'
         end